+++ python/assassyn/ir/expr/__init__.py
@@ -8,4 +8,5 @@ from .intrinsic import mem_write, send_read_request, send_write_request
 from .intrinsic import has_mem_resp, mem_resp, use_dram
 from .call import Bind, AsyncCall, FIFOPush
 from .comm import concat
+from .array import ArrayRead, ArrayWrite
 from . import comm
