+++ python/assassyn/ir/expr/expr.py
@@ -123,6 +123,7 @@ class Expr(Value):
         '''If this operation has a return value'''
         # pylint: disable=import-outside-toplevel
         from .intrinsic import PureIntrinsic
+        from .array import ArrayRead
         valued = (
             PureIntrinsic,
             FIFOPop,
@@ -163,81 +164,6 @@ class FIFOPop(Expr):
     def __getattr__(self, name):
         return self.dtype.attributize(self, name)
 
-
-class ArrayWrite(Expr):
-    '''The class for array write operation, where arr[idx] = val'''
-
-    ARRAY_WRITE = 401
-
-    def __init__(self, arr, idx: Value, val: Value):
-        super().__init__(ArrayWrite.ARRAY_WRITE, [arr, idx, val])
-
-    @property
-    def array(self) -> Array:
-        '''Get the array to write to'''
-        return self._operands[0]
-
-    @property
-    def idx(self) -> Value:
-        '''Get the index to write at'''
-        return self._operands[1]
-
-    @property
-    def val(self) -> Value:
-        '''Get the value to write'''
-        return self._operands[2]
-
-    def __repr__(self):
-        return f'{self.array.as_operand()}[{self.idx.as_operand()}] = {self.val.as_operand()}'
-
-
-class ArrayRead(Expr):
-    '''The class for array read operation, where arr[idx] as a right value'''
-
-    ARRAY_READ = 400
-
-    def __init__(self, arr: Array, idx: Value):
-        # pylint: disable=import-outside-toplevel
-        from ..array import Array
-        assert isinstance(arr, Array), f'{type(arr)} is not an Array!'
-        assert isinstance(idx, Value), f'{type(idx)} is not a Value!'
-        super().__init__(ArrayRead.ARRAY_READ, [arr, idx])
-
-    @property
-    def array(self) -> Array:
-        '''Get the array to read from'''
-        return self._operands[0]
-
-    @property
-    def idx(self) -> Value:
-        '''Get the index to read at'''
-        return self._operands[1]
-
-    @property
-    def dtype(self) -> DType:
-        '''Get the data type of the read value'''
-        return self.array.scalar_ty
-
-    def __repr__(self):
-        return f'{self.as_operand()} = {self.array.as_operand()}[{self.idx.as_operand()}]'
-
-    def __getattr__(self, name):
-        return self.dtype.attributize(self, name)
-
-    def __le__(self, value):
-        '''
-        Handle the <= operator for array writes.
-        '''
-        from ...builder import Singleton
-        from ..dtype import RecordValue
-
-        assert isinstance(value, (Value, RecordValue)), \
-            f"Value must be Value or RecordValue, got {type(value)}"
-
-        current_module = Singleton.builder.current_module
-
-        write_port = self.array & current_module
-        return write_port._create_write(self.idx.value, value)
 
 class Log(Expr):
     '''The class for log operation. NOTE: This operation is just like verilog $display, which is
